mypo-46 refactoring optimizer
