+++ mypo/__init__.py
@@ -5,7 +5,6 @@ from .common import calc_capital_gain_tax, calc_fee, calc_income_gain_tax
 from .loader import Loader
 from .market import Market
 from .model_selection import split_n_periods
-from .optimizer import MinimumVarianceOptimizer
 from .reporter import Reporter
 from .runner import Runner
 from .settings import Settings
