+++ mypo/optimizer/__init__.py
@@ -1,0 +1,3 @@
+# flake8: noqa
+
+from .optimizer import MinimumVarianceOptimizer
